feat: homepage redesign + rename games to challenges
- Add Two Paths section (Practice Tests + Challenge Modes cards)
- Add dismissible announcement banner for Challenge Modes
- Update hero stats: cert tracks → challenge modes
- Rename /games → /challenges routes
- Rename components/games → components/challenges
- Rename GameCard → ChallengeCard, GameSidebar → ChallengeSidebar
- Rename GameResult → ChallengeResult, GameType → ChallengeType
- Rename Games i18n namespace → Challenges
- Update all imports, hrefs, comments, and i18n keys
