+++ app/src/hooks/useTimeTrialMode.ts
@@ -14,7 +14,7 @@
 
 import { useState, useCallback, useEffect, useRef } from "react";
 import type { Question } from "@/types/quiz";
-import type { GameResult } from "@/types/games";
+import type { ChallengeResult } from "@/types/challenges";
 import { shuffle } from "@/lib/utils";
 
 type Phase = "loading" | "playing" | "feedback" | "wrong_review" | "paused" | "game_over";
@@ -279,7 +279,7 @@ export function useTimeTrialMode(allQuestions: Question[]) {
     }
   }, [state.phase, advanceToNext]);
 
-  const result: GameResult | null =
+  const result: ChallengeResult | null =
     state.phase === "game_over"
       ? {
           correct: state.correct,
